lnwallet: move txout serialization out of lnwire
